Trim 338 clippy warnings to zero (one external)
- Fix two FK-failing thread_ui_state tests by seeding an accounts row.
- Mechanical cleanups: collapsible_if -> let-chains; redundant_closure ->
  method refs; unnecessary_map_or -> is_none_or/is_some_and/is_ok_and;
  drop useless i64::from / .into() / format!() calls; small one-offs
  (manual_div_ceil, manual_is_multiple_of, manual_contains, derivable
  enum Defaults, etc).
- Module-level allow(dead_code) on genuine library modules (icon, font,
  layout, widgets, emoji_picker, undoable, theme).
- Per-item allow(dead_code) on intentional in-progress scaffolding
  (signature drag-reorder, status-bar staleness pruning, JMAP delegated
  sync, Graph contact sync, etc).
- allow(too_many_arguments) / allow(type_complexity) with TODO(refactor)
  notes on 19 db/UI sites; backlog tracked in TODO.md so the signal isn't
  lost.
- allow(should_implement_trait) on 4 from_str/next helpers that return
  Option / branded tokens.

Coded by an LLM.

+++ crates/graph/src/sync/persistence.rs
@@ -231,7 +231,7 @@ fn upsert_messages(
                 date: b.date,
                 is_read: b.is_read,
                 is_starred: b.is_starred,
-                raw_size: Some(i64::from(b.raw_size)),
+                raw_size: Some(b.raw_size),
                 internal_date: Some(b.internal_date),
                 list_unsubscribe: b.list_unsubscribe.clone(),
                 list_unsubscribe_post: b.list_unsubscribe_post.clone(),
